+++ dd-java-agent/instrumentation/lettuce-5/src/test/groovy/Lettuce5AsyncClientTest.groovy
@@ -28,76 +28,10 @@ import java.util.function.BiFunction
 import java.util.function.Consumer
 import java.util.function.Function
 
+import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
 import static datadog.trace.instrumentation.lettuce5.LettuceInstrumentationUtil.AGENT_CRASHING_COMMAND_PREFIX
 
-abstract class Lettuce5AsyncClientTest extends VersionedNamingTestBase {
-  public static final int DB_INDEX = 0
-  // Disable autoreconnect so we do not get stray traces popping up on server shutdown
-  public static final ClientOptions CLIENT_OPTIONS = ClientOptions.builder().autoReconnect(false).build()
-
-  @Shared
-  int port
-  @Shared
-  int incorrectPort
-  @Shared
-  String dbAddr
-  @Shared
-  String dbAddrNonExistent
-  @Shared
-  String dbUriNonExistent
-  @Shared
-  String embeddedDbUri
-
-  @Shared
-  RedisContainer redisServer = new RedisContainer(RedisContainer.DEFAULT_IMAGE_NAME)
-  .waitingFor(Wait.forListeningPort())
-
-  @Shared
-  Map<String, String> testHashMap = [
-    firstname: "John",
-    lastname : "Doe",
-    age      : "53"
-  ]
-
-  RedisClient redisClient
-  StatefulConnection connection
-  RedisAsyncCommands<String, ?> asyncCommands
-  RedisCommands<String, ?> syncCommands
-
-  def setupSpec() {
-    redisServer.start()
-    println "Using redis: $redisServer.redisURI"
-    port = redisServer.firstMappedPort
-    incorrectPort = PortUtils.randomOpenPort()
-    dbAddr = redisServer.getHost() + ":" + port + "/" + DB_INDEX
-    dbAddrNonExistent = redisServer.getHost() + ":" + incorrectPort + "/" + DB_INDEX
-    dbUriNonExistent = "redis://" + dbAddrNonExistent
-    embeddedDbUri = "redis://" + dbAddr
-  }
-
-  def setup() {
-    redisClient = RedisClient.create(embeddedDbUri)
-    redisClient.setOptions(CLIENT_OPTIONS)
-
-    connection = redisClient.connect()
-    asyncCommands = connection.async()
-    syncCommands = connection.sync()
-
-    syncCommands.set("TESTKEY", "TESTVAL")
-
-    // 1 set + 1 connect trace
-    TEST_WRITER.waitForTraces(2)
-    TEST_WRITER.clear()
-  }
-
-  def cleanup() {
-    connection.close()
-  }
-
-  def cleanupSpec() {
-    redisServer.stop()
-  }
-
+abstract class Lettuce5AsyncClientTest extends Lettuce5ClientTestBase {
   def "connect using get on ConnectionFuture"() {
     setup:
     RedisClient testConnectionClient = RedisClient.create(embeddedDbUri)
@@ -582,7 +516,7 @@ abstract class Lettuce5AsyncClientTest extends VersionedNamingTestBase {
   }
 }
 
-class Lettuce5SyncClientV0Test extends Lettuce5AsyncClientTest {
+class Lettuce5AsyncClientV0Test extends Lettuce5AsyncClientTest {
 
   @Override
   int version() {
@@ -600,7 +534,7 @@ class Lettuce5SyncClientV0Test extends Lettuce5AsyncClientTest {
   }
 }
 
-class Lettuce5SyncClientV1ForkedTest extends Lettuce5AsyncClientTest {
+class Lettuce5AsyncClientV1ForkedTest extends Lettuce5AsyncClientTest {
 
   @Override
   int version() {
@@ -619,7 +553,7 @@ class Lettuce5SyncClientV1ForkedTest extends Lettuce5AsyncClientTest {
 }
 
 
-class Lettuce5AsyncProfilingForkedTest extends Lettuce5AsyncClientTest {
+class Lettuce5AsyncClientProfilingForkedTest extends Lettuce5AsyncClientTest {
 
   @Override
   protected void configurePreAgent() {
