+++ dd-java-agent/instrumentation/lettuce-5/src/test/groovy/Lettuce5ClientTestBase.groovy
@@ -1,0 +1,78 @@
+import com.redis.testcontainers.RedisContainer
+import datadog.trace.agent.test.naming.VersionedNamingTestBase
+import datadog.trace.agent.test.utils.PortUtils
+import io.lettuce.core.ClientOptions
+import io.lettuce.core.RedisClient
+import io.lettuce.core.api.StatefulRedisConnection
+import io.lettuce.core.api.async.RedisAsyncCommands
+import io.lettuce.core.api.reactive.RedisReactiveCommands
+import io.lettuce.core.api.sync.RedisCommands
+import org.testcontainers.containers.wait.strategy.Wait
+import org.testcontainers.utility.DockerImageName
+import spock.util.concurrent.PollingConditions
+
+import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
+
+abstract class Lettuce5ClientTestBase extends VersionedNamingTestBase {
+  public static final int DB_INDEX = 0
+  // Disable autoreconnect so we do not get stray traces popping up on server shutdown
+  public static final ClientOptions CLIENT_OPTIONS = ClientOptions.builder().autoReconnect(false).build()
+
+  public static final Map<String, String> testHashMap = [
+    firstname: "John",
+    lastname : "Doe",
+    age      : "53"
+  ]
+
+  int port
+  int incorrectPort
+  String dbAddr
+  String dbAddrNonExistent
+  String dbUriNonExistent
+  String embeddedDbUri
+
+  RedisContainer redisServer = new RedisContainer(DockerImageName.parse("redis:6.2.6"))
+    .waitingFor(Wait.forListeningPort())
+
+  RedisClient redisClient
+  StatefulRedisConnection connection
+  RedisReactiveCommands<String, ?> reactiveCommands
+  RedisAsyncCommands<String, ?> asyncCommands
+  RedisCommands<String, ?> syncCommands
+
+  def setup() {
+    redisServer.start()
+    println "Using redis: $redisServer.redisURI"
+
+    port = redisServer.firstMappedPort
+    incorrectPort = PortUtils.randomOpenPort()
+    dbAddr = redisServer.getHost() + ":" + port + "/" + DB_INDEX
+    dbAddrNonExistent = redisServer.getHost() + ":" + incorrectPort + "/" + DB_INDEX
+    dbUriNonExistent = "redis://" + dbAddrNonExistent
+    embeddedDbUri = "redis://" + dbAddr
+
+    redisClient = RedisClient.create(embeddedDbUri)
+    redisClient.setOptions(CLIENT_OPTIONS)
+
+    runUnderTrace("setup") {
+      new PollingConditions(delay: 3, timeout: 15).eventually {
+        (connection = redisClient.connect()) != null
+      }
+      reactiveCommands = connection.reactive()
+      asyncCommands = connection.async()
+      syncCommands = connection.sync()
+
+      syncCommands.set("TESTKEY", "TESTVAL")
+      syncCommands.hmset("TESTHM", testHashMap)
+    }
+
+    TEST_WRITER.waitForTraces(1)
+    TEST_WRITER.clear()
+  }
+
+  def cleanup() {
+    connection.close()
+    redisClient.shutdown()
+    redisServer.stop()
+  }
+}
